# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LUMINA social network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/search:
    get:
      operationId: searchUsers
      tags: [users]
      summary: Search users by username or display name
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/suggestions:
    get:
      operationId: getUserSuggestions
      tags: [users]
      summary: Get suggested users to follow
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Suggestions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile with stats
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{username}/posts:
    get:
      operationId: listUserPosts
      tags: [users]
      summary: Get a user's posts
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedPost"
  /users/{username}/followers:
    get:
      operationId: getUserFollowers
      tags: [users]
      summary: Get a user's followers
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Followers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{username}/following:
    get:
      operationId: getUserFollowing
      tags: [users]
      summary: Get users that a user follows
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Following
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /users/{username}/follow:
    post:
      operationId: toggleFollow
      tags: [users]
      summary: Toggle follow on a user
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Follow toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowResult"

  # ── STORIES ─────────────────────────────────────────────────────────────────
  /stories:
    get:
      operationId: getStories
      tags: [stories]
      summary: Get active stories from followed users grouped by user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Story groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StoryGroup"

    post:
      operationId: createStory
      tags: [stories]
      summary: Create a story
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Story"
  /stories/{id}/view:
    post:
      operationId: viewStory
      tags: [stories]
      summary: Record a story view
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: View recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewResult"
  /stories/{id}/viewers:
    get:
      operationId: getStoryViewers
      tags: [stories]
      summary: Get who viewed a story (owner only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Viewers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSummary"
  /stories/{id}:
    delete:
      operationId: deleteStory
      tags: [stories]
      summary: Delete a story
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    UserSummary:
      type: object
      required: [id, username, displayName, avatarUrl, verified]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        verified:
          type: boolean
        bio:
          type: string
        isFollowing:
          type: boolean
        followerCount:
          type: integer
    UserProfile:
      type: object
      required: [id, username, displayName, avatarUrl, coverUrl, bio, website, location, verified, createdAt, postCount, followerCount, followingCount, isFollowing]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: string
        coverUrl:
          type: string
        bio:
          type: string
        website:
          type: string
        location:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
        postCount:
          type: integer
        followerCount:
          type: integer
        followingCount:
          type: integer
        isFollowing:
          type: boolean

    # ── Post schemas ─────────────────────────────────────────────────────────
...
    FeedPost:
      type: object
      required: [id, content, imageUrl, imageUrl2, postType, views, createdAt, author, likeCount, commentCount, liked, saved, topReactions, previewComments]
      properties:
        id:
          type: integer
        content:
          type: string
        imageUrl:
          type: string
        imageUrl2:
          type: string
        postType:
          type: string
        views:
          type: integer
        scheduledAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/UserSummary"
        likeCount:
          type: integer
        commentCount:
          type: integer
        liked:
          type: boolean
        saved:
          type: boolean
        myReaction:
          type: ["string", "null"]
        topReactions:
          type: array
          items:
            $ref: "#/components/schemas/ReactionCount"
        previewComments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
...
    ReactionCount:
      type: object
      required: [reaction, count]
      properties:
        reaction:
          type: string
        count:
          type: integer
...
    Comment:
      type: object
      required: [id, postId, content, createdAt, author, likeCount, liked]
      properties:
        id:
          type: integer
        postId:
          type: integer
        content:
          type: string
        parentCommentId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        author:
          $ref: "#/components/schemas/UserSummary"
        likeCount:
          type: integer
        liked:
          type: boolean

    # ── Follow / Reaction schemas ────────────────────────────────────────────
    FollowResult:
      type: object
      required: [following, count]
      properties:
        following:
          type: boolean
        count:
          type: integer
...
    StoryInput:
      type: object
      properties:
        mediaUrl:
          type: ["string", "null"]
        mediaType:
          type: string
          enum: [image, video, text]
          default: text
        caption:
          type: string
        bgColor:
          type: string
          default: "#7c6aff"
        textContent:
          type: string
    Story:
      type: object
      required: [id, userId, mediaUrl, mediaType, caption, bgColor, textContent, views, expiresAt, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        mediaUrl:
          type: string
        mediaType:
          type: string
        caption:
          type: string
        bgColor:
          type: string
        textContent:
          type: string
        views:
          type: integer
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    StoryGroup:
      type: object
      required: [user, stories, hasUnviewed]
      properties:
        user:
          $ref: "#/components/schemas/UserSummary"
        stories:
          type: array
          items:
            $ref: "#/components/schemas/Story"
        hasUnviewed:
          type: boolean
    ViewResult:
      type: object
      required: [viewCount]
      properties:
        viewCount:
          type: integer

    # ── Notification schemas ─────────────────────────────────────────────────
...