# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LUMINA social network API
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: number;
  userId: number;
  mediaUrl: string;
  mediaType: string;
  caption: string;
  bgColor: string;
  textContent: string;
  views: number;
  expiresAt: Date;
  createdAt: Date;
}